requirements.txt
# CAN
python-can

# LiDAR
# git+https://github.com/simondlevy/BreezyLidar.git#egg=BreezyLidar&subdirectory=python
BreezyLidar

# web services
flask
socketio
python-engineio
python-socketio
websockets
websocket-client
gevent==22.8.0
gevent-websocket
eventlet
pymitter
